refactor(acrawl-cli): split app.rs into app/ module directory
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
